work in progress
Former-commit-id: 7e9ac6e71aff34f4e5d0ee797666f31f4bae6a79
Former-commit-id: c683bf87f23f7c66222bb1483a430fd74469f284

+++ src/ttf.h
@@ -13,15 +13,15 @@ void ttf_free(fontp);
 void ttf_text_size(fontp, const char *text,
                    uint32_t *w, uint32_t *h,
                    enum_fmt *fmt,
-                   float scaling,
-                   float advance,
+                   double scaling,
+                   double advance,
                    boolean fixed_width);
-void ttf_putc(fontp, int32_t c, int32_t x, int32_t y, float scaling);
-void ttf_puts(fontp, const char *text, int32_t x, int32_t y, float scaling,
-              float advance,
+void ttf_putc(fontp, int32_t c, int32_t x, int32_t y, double scaling);
+void ttf_puts(fontp, const char *text, int32_t x, int32_t y, double scaling,
+              double advance,
               boolean fixed_width);
-void ttf_puts_no_fmt(fontp, const char *text, int32_t x, int32_t y, float scaling,
-              float advance,
+void ttf_puts_no_fmt(fontp, const char *text, int32_t x, int32_t y, double scaling,
+              double advance,
               boolean fixed_width);
 void ttf2tga(void);
 fontp ttf_read_tga(char *name, int32_t pointsize);
