+++ src/wid_text_input.h
@@ -4,5 +4,5 @@
  * See the README file for license.
  */
 
-widp wid_text_input(const char *title, float x, float y, int32_t args, ...);
-widp wid_large_text_input(const char *title, float x, float y, int32_t args, ...);
+widp wid_text_input(const char *title, double x, double y, int32_t args, ...);
+widp wid_large_text_input(const char *title, double x, double y, int32_t args, ...);
